first commit
This is an excel file that can calculate the quality, amount, and cost of agricultural lime. You can analyze up to five sources simultaneously. Please be sure to select the number of lime sources from the drop-down list before inserting values.  You can use the simple "Data Entry Form" in the Dashboard. The other form serves as a make up table.